tree_wide: rename s3 directory to ...
... cloud_storage_clients

The s3 build target has also been renamed to cloud_storage_clients.
