+++ src/v/CMakeLists.txt
@@ -27,7 +27,7 @@ add_subdirectory(kafka)
 add_subdirectory(reflection)
 add_subdirectory(pandaproxy)
 add_subdirectory(http)
-add_subdirectory(s3)
+add_subdirectory(cloud_storage_clients)
 add_subdirectory(archival)
 add_subdirectory(security)
 add_subdirectory(serde)
