+++ src/v/archival/ntp_archiver_service.h
@@ -15,11 +15,11 @@
 #include "cloud_storage/partition_manifest.h"
 #include "cloud_storage/remote.h"
 #include "cloud_storage/types.h"
+#include "cloud_storage_clients/client.h"
 #include "cluster/fwd.h"
 #include "cluster/partition.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
-#include "s3/client.h"
 #include "storage/fwd.h"
 #include "storage/segment.h"
 #include "utils/retry_chain_node.h"
