+++ src/v/archival/tests/service_fixture.cc
@@ -13,11 +13,11 @@
 #include "archival/types.h"
 #include "bytes/iobuf.h"
 #include "bytes/iobuf_parser.h"
+#include "cloud_storage_clients/configuration.h"
 #include "cluster/archival_metadata_stm.h"
 #include "cluster/members_table.h"
 #include "model/tests/random_batch.h"
 #include "random/generators.h"
-#include "s3/configuration.h"
 #include "seastarx.h"
 #include "storage/directories.h"
 #include "storage/disk_log_impl.h"
