+++ src/v/cloud_storage/remote.h
@@ -15,10 +15,10 @@
 #include "cloud_storage/fwd.h"
 #include "cloud_storage/probe.h"
 #include "cloud_storage/types.h"
+#include "cloud_storage_clients/client.h"
+#include "cloud_storage_clients/client_pool.h"
 #include "model/metadata.h"
 #include "random/simple_time_jitter.h"
-#include "s3/client.h"
-#include "s3/client_pool.h"
 #include "storage/segment_reader.h"
 #include "utils/retry_chain_node.h"
