+++ src/v/cloud_storage/remote_partition.h
@@ -17,9 +17,9 @@
 #include "cloud_storage/remote_segment.h"
 #include "cloud_storage/segment_state.h"
 #include "cloud_storage/types.h"
+#include "cloud_storage_clients/types.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
-#include "s3/types.h"
 #include "storage/ntp_config.h"
 #include "storage/translating_reader.h"
 #include "storage/types.h"
