+++ src/v/cloud_storage/tests/remote_partition_test.cc
@@ -20,12 +20,12 @@
 #include "cloud_storage/tests/common_def.h"
 #include "cloud_storage/tests/s3_imposter.h"
 #include "cloud_storage/types.h"
+#include "cloud_storage_clients/types.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/record.h"
 #include "model/record_batch_types.h"
 #include "model/timeout_clock.h"
-#include "s3/types.h"
 #include "seastarx.h"
 #include "storage/log.h"
 #include "storage/log_manager.h"
