+++ src/v/cloud_storage_clients/configuration.h
@@ -11,10 +11,10 @@
 #pragma once
 
 #include "cloud_roles/types.h"
+#include "cloud_storage_clients/client_probe.h"
+#include "cloud_storage_clients/types.h"
 #include "net/transport.h"
 #include "net/types.h"
-#include "s3/client_probe.h"
-#include "s3/types.h"
 
 namespace cloud_storage_clients {
