+++ src/v/cloud_storage_clients/error.h
