+++ src/v/cloud_storage_clients/logger.h
