+++ src/v/cloud_storage_clients/test_client/CMakeLists.txt
@@ -1,5 +1,5 @@
 # client
 add_executable(s3_test_client s3_test_client_main.cc)
-target_link_libraries(s3_test_client PUBLIC v::model v::net v::http v::s3 v::cloud_roles)
+target_link_libraries(s3_test_client PUBLIC v::model v::net v::http v::cloud_storage_clients v::cloud_roles)
 set_property(TARGET s3_test_client PROPERTY POSITION_INDEPENDENT_CODE ON)
