+++ src/v/cloud_storage_clients/tests/CMakeLists.txt
@@ -5,7 +5,7 @@ rp_test(
   SOURCES
     s3_client_test.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
-  LIBRARIES v::seastar_testing_main Boost::unit_test_framework v::http v::s3 v::cloud_roles
+  LIBRARIES v::seastar_testing_main Boost::unit_test_framework v::http v::cloud_storage_clients v::cloud_roles
   ARGS "-- -c 1"
   LABELS s3
 )
