+++ src/v/cloud_storage_clients/tests/s3_client_test.cc
@@ -12,12 +12,12 @@
 #include "bytes/iobuf_parser.h"
 #include "bytes/iostream.h"
 #include "cloud_roles/signature.h"
+#include "cloud_storage_clients/client_pool.h"
+#include "cloud_storage_clients/error.h"
+#include "cloud_storage_clients/s3_client.h"
 #include "net/dns.h"
 #include "net/types.h"
 #include "net/unresolved_address.h"
-#include "s3/client_pool.h"
-#include "s3/error.h"
-#include "s3/s3_client.h"
 #include "seastarx.h"
 
 #include <seastar/core/future.hh>
