+++ src/v/cloud_storage_clients/types.h
