+++ src/v/cluster/remote_topic_configuration_source.cc
@@ -14,10 +14,10 @@
 #include "cloud_storage/remote.h"
 #include "cloud_storage/topic_manifest.h"
 #include "cloud_storage/types.h"
+#include "cloud_storage_clients/configuration.h"
 #include "cluster/logger.h"
 #include "cluster/types.h"
 #include "config/configuration.h"
-#include "s3/configuration.h"
 
 namespace cluster {
